move Crystal to own dir
