+++ packages/quicktype-core/src/language/Crystal/CrystalRenderer.ts
@@ -1,234 +1,14 @@
-import { anyTypeIssueAnnotation, nullTypeIssueAnnotation } from "../Annotation";
-import { ConvenienceRenderer, type ForbiddenWordsInfo } from "../ConvenienceRenderer";
-import { type Name, type Namer, funPrefixNamer } from "../Naming";
-import { type RenderContext } from "../Renderer";
-import { type Option } from "../RendererOptions";
-import { type Sourcelike, maybeAnnotated } from "../Source";
-import {
-    allLowerWordStyle,
-    combineWords,
-    escapeNonPrintableMapper,
-    firstUpperWordStyle,
-    intToHex,
-    isAscii,
-    isLetterOrUnderscore,
-    isLetterOrUnderscoreOrDigit,
-    isPrintable,
-    legalizeCharacters,
-    splitIntoWords,
-    utf32ConcatMap
-} from "../support/Strings";
-import { TargetLanguage } from "../TargetLanguage";
-import { type ClassType, type EnumType, type Type, type UnionType } from "../Type";
-import { type FixMeOptionsAnyType } from "../types";
-import { matchType, nullableFromUnion, removeNullFromUnion } from "../TypeUtils";
+import { anyTypeIssueAnnotation, nullTypeIssueAnnotation } from "../../Annotation";
+import { ConvenienceRenderer, type ForbiddenWordsInfo } from "../../ConvenienceRenderer";
+import { type Name, type Namer } from "../../Naming";
+import { type RenderContext } from "../../Renderer";
+import { type Sourcelike, maybeAnnotated } from "../../Source";
+import { type TargetLanguage } from "../../TargetLanguage";
+import { type ClassType, type EnumType, type Type, type UnionType } from "../../Type";
+import { matchType, nullableFromUnion, removeNullFromUnion } from "../../TypeUtils";
 
-export class CrystalTargetLanguage extends TargetLanguage {
-    protected makeRenderer(renderContext: RenderContext): CrystalRenderer {
-        return new CrystalRenderer(this, renderContext);
-    }
-
-    public constructor() {
-        super("Crystal", ["crystal", "cr", "crystallang"], "cr");
-    }
-
-    protected get defaultIndentation(): string {
-        return "  ";
-    }
-
-    protected getOptions(): Array<Option<FixMeOptionsAnyType>> {
-        return [];
-    }
-}
-
-const keywords = [
-    "Any",
-    "Array",
-    "Atomic",
-    "Bool",
-    "Channel",
-    "Char",
-    "Class",
-    "Enum",
-    "Enumerable",
-    "Event",
-    "Extern",
-    "Exception",
-    "File",
-    "Float",
-    "Float32",
-    "Float64",
-    "GC",
-    "GZip",
-    "Hash",
-    "HTML",
-    "HTTP",
-    "Int",
-    "Int128",
-    "Int16",
-    "Int32",
-    "Int64",
-    "Int8",
-    "Iterable",
-    "Link",
-    "Logger",
-    "Math",
-    "Mutex",
-    "Nil",
-    "Number",
-    "JSON",
-    "IO",
-    "Object",
-    "Pointer",
-    "Proc",
-    "Process",
-    "Range",
-    "Random",
-    "Regex",
-    "Reference",
-    "Set",
-    "Signal",
-    "Slice",
-    "Spec",
-    "StaticArray",
-    "String",
-    "Struct",
-    "Symbol",
-    "System",
-    "TCPServer",
-    "TCPSocket",
-    "Socket",
-    "Tempfile",
-    "Termios",
-    "Time",
-    "Tuple",
-    "ThreadLocal",
-    "UDPSocket",
-    "UInt128",
-    "UInt16",
-    "UInt32",
-    "UInt64",
-    "UInt8",
-    "Union",
-    "UNIXServer",
-    "UNIXSocket",
-    "UUID",
-    "URI",
-    "VaList",
-    "Value",
-    "Void",
-    "WeakRef",
-    "XML",
-    "YAML",
-    "Zip",
-    "Zlib",
-    "abstract",
-    "alias",
-    "as",
-    "as?",
-    "asm",
-    "begin",
-    "break",
-    "case",
-    "class",
-    "def",
-    "do",
-    "else",
-    "elsif",
-    "end",
-    "ensure",
-    "enum",
-    "extend",
-    "false",
-    "for",
-    "fun",
-    "if",
-    "in",
-    "include",
-    "instance_sizeof",
-    "is_a?",
-    "lib",
-    "macro",
-    "module",
-    "next",
-    "nil",
-    "nil?",
-    "of",
-    "out",
-    "pointerof",
-    "private",
-    "protected",
-    "require",
-    "rescue",
-    "return",
-    "select",
-    "self",
-    "sizeof",
-    "struct",
-    "super",
-    "then",
-    "true",
-    "type",
-    "typeof",
-    "uninitialized",
-    "union",
-    "unless",
-    "until",
-    "when",
-    "while",
-    "with",
-    "yield"
-];
-
-function isAsciiLetterOrUnderscoreOrDigit(codePoint: number): boolean {
-    if (!isAscii(codePoint)) {
-        return false;
-    }
-
-    return isLetterOrUnderscoreOrDigit(codePoint);
-}
-
-function isAsciiLetterOrUnderscore(codePoint: number): boolean {
-    if (!isAscii(codePoint)) {
-        return false;
-    }
-
-    return isLetterOrUnderscore(codePoint);
-}
-
-const legalizeName = legalizeCharacters(isAsciiLetterOrUnderscoreOrDigit);
-
-function crystalStyle(original: string, isSnakeCase: boolean): string {
-    const words = splitIntoWords(original);
-
-    const wordStyle = isSnakeCase ? allLowerWordStyle : firstUpperWordStyle;
-
-    const combined = combineWords(
-        words,
-        legalizeName,
-        wordStyle,
-        wordStyle,
-        wordStyle,
-        wordStyle,
-        isSnakeCase ? "_" : "",
-        isAsciiLetterOrUnderscore
-    );
-
-    return combined === "_" ? "_underscore" : combined;
-}
-
-const snakeNamingFunction = funPrefixNamer("default", (original: string) => crystalStyle(original, true));
-const camelNamingFunction = funPrefixNamer("camel", (original: string) => crystalStyle(original, false));
-
-function standardUnicodeCrystalEscape(codePoint: number): string {
-    if (codePoint <= 0xffff) {
-        return "\\u{" + intToHex(codePoint, 4) + "}";
-    } else {
-        return "\\u{" + intToHex(codePoint, 6) + "}";
-    }
-}
-
-const crystalStringEscape = utf32ConcatMap(escapeNonPrintableMapper(isPrintable, standardUnicodeCrystalEscape));
+import { keywords } from "./constants";
+import { camelNamingFunction, crystalStringEscape, snakeNamingFunction } from "./utils";
 
 export class CrystalRenderer extends ConvenienceRenderer {
     public constructor(targetLanguage: TargetLanguage, renderContext: RenderContext) {
@@ -251,7 +31,7 @@ export class CrystalRenderer extends ConvenienceRenderer {
         return camelNamingFunction;
     }
 
-    protected forbiddenNamesForGlobalNamespace(): string[] {
+    protected forbiddenNamesForGlobalNamespace(): readonly string[] {
         return keywords;
     }
