+++ packages/quicktype-core/src/language/Crystal/constants.ts
@@ -1,0 +1,138 @@
+export const keywords = [
+    "Any",
+    "Array",
+    "Atomic",
+    "Bool",
+    "Channel",
+    "Char",
+    "Class",
+    "Enum",
+    "Enumerable",
+    "Event",
+    "Extern",
+    "Exception",
+    "File",
+    "Float",
+    "Float32",
+    "Float64",
+    "GC",
+    "GZip",
+    "Hash",
+    "HTML",
+    "HTTP",
+    "Int",
+    "Int128",
+    "Int16",
+    "Int32",
+    "Int64",
+    "Int8",
+    "Iterable",
+    "Link",
+    "Logger",
+    "Math",
+    "Mutex",
+    "Nil",
+    "Number",
+    "JSON",
+    "IO",
+    "Object",
+    "Pointer",
+    "Proc",
+    "Process",
+    "Range",
+    "Random",
+    "Regex",
+    "Reference",
+    "Set",
+    "Signal",
+    "Slice",
+    "Spec",
+    "StaticArray",
+    "String",
+    "Struct",
+    "Symbol",
+    "System",
+    "TCPServer",
+    "TCPSocket",
+    "Socket",
+    "Tempfile",
+    "Termios",
+    "Time",
+    "Tuple",
+    "ThreadLocal",
+    "UDPSocket",
+    "UInt128",
+    "UInt16",
+    "UInt32",
+    "UInt64",
+    "UInt8",
+    "Union",
+    "UNIXServer",
+    "UNIXSocket",
+    "UUID",
+    "URI",
+    "VaList",
+    "Value",
+    "Void",
+    "WeakRef",
+    "XML",
+    "YAML",
+    "Zip",
+    "Zlib",
+    "abstract",
+    "alias",
+    "as",
+    "as?",
+    "asm",
+    "begin",
+    "break",
+    "case",
+    "class",
+    "def",
+    "do",
+    "else",
+    "elsif",
+    "end",
+    "ensure",
+    "enum",
+    "extend",
+    "false",
+    "for",
+    "fun",
+    "if",
+    "in",
+    "include",
+    "instance_sizeof",
+    "is_a?",
+    "lib",
+    "macro",
+    "module",
+    "next",
+    "nil",
+    "nil?",
+    "of",
+    "out",
+    "pointerof",
+    "private",
+    "protected",
+    "require",
+    "rescue",
+    "return",
+    "select",
+    "self",
+    "sizeof",
+    "struct",
+    "super",
+    "then",
+    "true",
+    "type",
+    "typeof",
+    "uninitialized",
+    "union",
+    "unless",
+    "until",
+    "when",
+    "while",
+    "with",
+    "yield"
+] as const;
