+++ packages/quicktype-core/src/language/Crystal/index.ts
@@ -1,0 +1,2 @@
+export { CrystalTargetLanguage } from "./language";
+export { CrystalRenderer } from "./CrystalRenderer";
